requirements.txt
django~=4.1.3
strawberry-graphql~=0.217.1
strawberry-graphql-django~=0.28.3